Consolidate LoginModule functionality in the abstract class. Also implement abort() properly.
